+++ Wire-iOS/Sources/UserInterface/Conversation/Content/Cells/ParticipantsCell/ParticipantsCellViewModel.swift
@@ -20,23 +20,11 @@
 enum ConversationActionType {
 
     case none, started(withName: String?), added(herself: Bool), removed, left, teamMemberLeave
-    // MOVE THIS OUT OF HERE
-    func formatKey(senderIsSelfUser: Bool) -> String {
-        switch self {
-        case .left: return localizationKey(with: "left", senderIsSelfUser: senderIsSelfUser)
-        case .added(herself: true): return "content.system.conversation.guest.joined"
-        case .added(herself: false): return localizationKey(with: "added", senderIsSelfUser: senderIsSelfUser)
-        case .removed: return localizationKey(with: "removed", senderIsSelfUser: senderIsSelfUser)
-        case .started(withName: .none), .none: return localizationKey(with: "started", senderIsSelfUser: senderIsSelfUser)
-        case .started(withName: .some): return "content.system.conversation.with_name.participants"
-        case .teamMemberLeave: return "content.system.conversation.team.member-leave"
-        }
-    }
     
     var involvesUsersOtherThanSender: Bool {
         switch self {
         case .left, .teamMemberLeave, .added(herself: true): return false
-        default: return true
+        default:                                             return true
         }
     }
 
@@ -50,13 +38,7 @@ enum ConversationActionType {
         
         return UIImage(for: icon, iconSize: .tiny, color: color)
     }
-    // THIS TOO
-    private func localizationKey(with pathComponent: String, senderIsSelfUser: Bool) -> String {
-        let senderPath = senderIsSelfUser ? "you" : "other"
-        return "content.system.conversation.\(senderPath).\(pathComponent)"
-    }
 }
-
 
 extension ZMConversationMessage {
     var actionType: ConversationActionType {
@@ -71,14 +53,18 @@ extension ZMConversationMessage {
     }
 }
 
-
 struct ParticipantsCellViewModel {
-
+    
+    private typealias NameList = ParticipantsStringFormatter.NameList
     static let showMoreLinkURL = NSURL(string: "action://show-all")!
     
     let font, boldFont, largeFont: UIFont?
     let textColor: UIColor?
     let message: ZMConversationMessage
+    
+    private var action: ConversationActionType {
+        return message.actionType
+    }
     
     private var maxShownUsers: Int {
         return isSelfIncludedInUsers ? 16 : 17
@@ -89,8 +75,7 @@ struct ParticipantsCellViewModel {
     }
     
     var showInviteButton: Bool {
-        guard case .started = message.actionType,
-            let conversation = message.conversation else { return false }
+        guard case .started = action, let conversation = message.conversation else { return false }
         return conversation.canManageAccess && conversation.allowGuests
     }
     
@@ -114,7 +99,7 @@ struct ParticipantsCellViewModel {
     /// The users represented by the collapsed link after being added to the
     /// conversation.
     var selectedUsers: [ZMUser] {
-        switch message.actionType {
+        switch action {
         case .added: return collapsedUsers
         default: return []
         }
@@ -128,7 +113,7 @@ struct ParticipantsCellViewModel {
     /// name.
     func sortedUsers() -> [ZMUser] {
         guard let sender = message.sender else { return [] }
-        guard message.actionType.involvesUsersOtherThanSender else { return [sender] }
+        guard action.involvesUsersOtherThanSender else { return [sender] }
         guard let systemMessage = message.systemMessageData else { return [] }
         return systemMessage.users.subtracting([sender]).sorted { name(for: $0) < name(for: $1) }
     }
@@ -148,260 +133,61 @@ struct ParticipantsCellViewModel {
         }
     }
     
+    private var nameList: NameList {
+        let userNames = shownUsers.map { self.name(for: $0) }
+        return NameList(names: userNames, collapsed: collapsedUsers.count, selfIncluded: isSelfIncludedInUsers)
+    }
+    
     /// The user will, depending on the context, be in a specific case within the
     /// sentence. This is important for localization of "you".
     private func grammaticalCase(for user: ZMUser) -> String {
-        if user == message.sender {
-            // sender is always the subject doing the action
-            return "nominative"
-        } else if case ConversationActionType.started = message.actionType {
-            // "sender started the conversation WITH ... user"
-            return "dative"
-        } else {
-            return "accusative"
-        }
+        // user is always the subject
+        if user == message.sender { return "nominative" }
+        // "started with ... user"
+        if case .started = action { return "dative" }
+        return "accusative"
     }
     
     // ------------------------------------------------------------
     
     func image() -> UIImage? {
-        return message.actionType.image(with: textColor)
+        return action.image(with: textColor)
     }
     
     func attributedHeading() -> NSAttributedString? {
         guard
-            case let .started(withName: conversationName?) = message.actionType,
+            case let .started(withName: conversationName?) = action,
             let sender = message.sender,
-            let font = font,
-            let boldFont = boldFont,
-            let largeFont = largeFont,
-            let textColor = textColor
+            let formatter = formatter(for: message)
             else { return nil }
-        
-        let formatter = ParticipantsStringFormatter(
-            message: message,
-            font: font,
-            boldFont: boldFont,
-            largeFont: largeFont,
-            textColor: textColor
-        )
         
         let senderName = name(for: sender).capitalized
         return formatter.heading(senderName: senderName, senderIsSelf: sender.isSelfUser, convName: conversationName)
     }
-    
+
     func attributedTitle() -> NSAttributedString? {
         guard
             let sender = message.sender,
-            let font = font,
-            let boldFont = boldFont,
-            let textColor = textColor
+            let formatter = formatter(for: message)
             else { return nil }
-        
-        // REFACTOR THIS, DON'T WANT TO PASS THE LARGE FONT HERE
-        let formatter = ParticipantsStringFormatter(
-            message: message,
-            font: font,
-            boldFont: boldFont,
-            largeFont: largeFont!,
-            textColor: textColor
-        )
         
         let senderName = name(for: sender).capitalized
         
-        if message.actionType.involvesUsersOtherThanSender {
-            let userNames = shownUsers.map { self.name(for: $0) }
-            let nameList = NameList(names: userNames, collapsed: collapsedUsers.count, selfIncluded: isSelfIncludedInUsers)
+        if action.involvesUsersOtherThanSender {
             return formatter.title(senderName: senderName, senderIsSelf: sender.isSelfUser, names: nameList)
         } else {
             return formatter.title(senderName: senderName, senderIsSelf: sender.isSelfUser)
         }
     }
-
-}
-
-
-private typealias Attributes = [NSAttributedStringKey : AnyObject]
-
-private class FormatSequence {
-    typealias SubstringAttrs = (substring: String, attrs: Attributes)
-    var string = String()
-    var componentAttributes = [SubstringAttrs]()
     
-    func append(_ component: String, with attrs: Attributes) {
-        string.append(component)
-        define(attrs, forComponent: component)
-    }
-    
-    func define(_ attrs: Attributes, forComponent string: String) {
-        componentAttributes.append(SubstringAttrs(string, attrs))
-    }
-    
-    func applyComponentAttributes(to attributedString: NSAttributedString) -> NSAttributedString {
-        let mutableCopy = NSMutableAttributedString(attributedString: attributedString)
-        componentAttributes.forEach {
-            mutableCopy.addAttributes($0.attrs, to: $0.substring)
-        }
-        return mutableCopy
-    }
-}
-
-
-private struct NameList {
-    let names: [String]
-    let collapsed: Int
-    let selfIncluded: Bool
-}
-
-private class ParticipantsStringFormatter {
-    
-    private let kYouStartedTheConversation = "content.system.conversation.with_name.title-you"
-    private let kXStartedTheConversation = "content.system.conversation.with_name.title"
-    private let kXOthers = "content.system.started_conversation.truncated_people.others"
-    private let kAndX = "content.system.started_conversation.truncated_people"
-    private let kWith = "content.system.conversation.with_name.participants"
-    private let kXAndY = "content.system.participants_1_other"
-    private let kCompleteTeam = "content.system.started_conversation.complete_team"
-    private let kCompleteTeamWithGuests = "content.system.started_conversation.complete_team.guests"
-    
-    private let font, boldFont, largeFont: UIFont
-    private let textColor: UIColor
-    
-    private let message: ZMConversationMessage
-    
-    private var normalAttributes: Attributes {
-        return [.font: font, .foregroundColor: textColor]
-    }
-    
-    private var boldAttributes: Attributes {
-        return [.font: boldFont, .foregroundColor: textColor]
-    }
-    
-    private var largeAttributes: Attributes {
-        return [.font: largeFont, .foregroundColor: textColor]
-    }
-    
-    private var linkAttributes: Attributes {
-        return [.link: ParticipantsCellViewModel.showMoreLinkURL]
-    }
-    
-    init(message: ZMConversationMessage, font: UIFont, boldFont: UIFont, largeFont: UIFont, textColor: UIColor) {
-        self.message = message
-        self.font = font
-        self.boldFont = boldFont
-        self.largeFont = largeFont
-        self.textColor = textColor
-    }
-    
-    func heading(senderName: String, senderIsSelf: Bool, convName: String) -> NSAttributedString {
-        // "... started the conversation"
-        var text: NSAttributedString
-        if senderIsSelf {
-            text = kYouStartedTheConversation.localized(args: senderName) && font
-        } else {
-            text = kXStartedTheConversation.localized(args: senderName) && font
-            text = text.adding(font: boldFont, to: senderName)
-        }
-        // "Italy Trip"
-        let title = convName.attributedString && largeFont
-        return [text, title].joined(separator: "\n".attributedString) && textColor && .lineSpacing(4)
-    }
-
-    // title involving only subject (sender)
-    func title(senderName: String, senderIsSelf: Bool) -> NSAttributedString? {
-        switch message.actionType {
-        case .left, .teamMemberLeave, .added(herself: true):
-            let formatKey = message.actionType.formatKey
-            let title = formatKey(senderIsSelf).localized(args: senderName) && font && textColor
-            return senderIsSelf ? title : title.adding(font: boldFont, to: senderName)
-        default: return nil
-        }
-    }
-    
-    // title involving subject (sender) and object(s) (users)
-    func title(senderName: String, senderIsSelf: Bool, names: NameList) -> NSAttributedString? {
-        let formatKey = message.actionType.formatKey
-        let namesFormat = nameListFormat(for: names)
-        
-        switch message.actionType {
-        case .removed, .added(herself: false), .started(withName: .none):
-            var title = formatKey(senderIsSelf).localized(args: senderName, namesFormat.string) && font && textColor
-            if !senderIsSelf { title = title.adding(font: boldFont, to: senderName) }
-            return namesFormat.applyComponentAttributes(to: title)
-            
-        case .started(withName: .some):
-            let title = "\(kWith.localized) \(namesFormat.string)" && font && textColor
-            // this could be refactored
-            return namesFormat.applyComponentAttributes(to: title)
-        default: return nil
-        }
-    }
-    
-    /// Returns a `FormatSequence` describing a list of names. The list is comprised
-    /// of usernames for shown users (complete with punctuation) and a count string
-    /// for collapsed users, if any.
-    /// E.g: "x, y, z, and 3 others"
-
-    private func nameListFormat(for nameList: NameList) -> FormatSequence {
-        // there must be some names
-        guard !nameList.names.isEmpty else { preconditionFailure() }
-
-        let result = FormatSequence()
-
-        // all team users added?
-        if let linkText = linkTextForWholeTeam() {
-            result.append(linkText, with: linkAttributes)
-            return result
-        }
-        
-        let attrsForLastName = nameList.selfIncluded ? normalAttributes : boldAttributes
-        let names = nameList.names
-        
-        switch names.count {
-        case 1:
-            // "x"
-            result.append(names.last!, with: attrsForLastName)
-        case 2:
-            // "x and y"
-            let part = kAndX.localized(args: names)
-            result.append(part, with: normalAttributes)
-            result.define(boldAttributes, forComponent: names.first!)
-            result.define(attrsForLastName, forComponent: names.last!)
-        default:
-            // "x, y, "
-            result.append(names.dropLast().map { $0 + ", " }.joined(), with: boldAttributes)
-            
-            if nameList.collapsed > 0 {
-                // "you/z, "
-                result.append(names.last! + ", ", with: attrsForLastName)
-                // "and X others
-                let linkText = kXOthers.localized(args: "\(nameList.collapsed)")
-                let linkPart = kAndX.localized(args: linkText)
-                result.append(linkPart, with: normalAttributes)
-                result.define(linkAttributes, forComponent: linkText)
-            } else {
-                // "and you/z"
-                let lastPart = kAndX.localized(args: names.last!)
-                result.append(lastPart, with: normalAttributes)
-                result.define(attrsForLastName, forComponent: names.last!)
-            }
-        }
-        
-        return result
-    }
-    
-    private func linkTextForWholeTeam() -> String? {
-        guard
-            let systemMessage = message as? ZMSystemMessage,
-            systemMessage.allTeamUsersAdded,
-            message.conversation?.canManageAccess ?? false
+    private func formatter(for message: ZMConversationMessage) -> ParticipantsStringFormatter? {
+        guard let font = font, let boldFont = boldFont,
+            let largeFont = largeFont, let textColor = textColor
             else { return nil }
         
-        
-        if systemMessage.numberOfGuestsAdded > 0 {
-            return kCompleteTeamWithGuests.localized(args: String(systemMessage.numberOfGuestsAdded))
-        } else {
-            return kCompleteTeam.localized
-        }
+        return ParticipantsStringFormatter(
+            message: message, font: font, boldFont: boldFont,
+            largeFont: largeFont, textColor: textColor
+        )
     }
 }
